refactor: entity 찾지 못할 시에 NotFoundException 을 상속하는 custom exception class 를 사용하도록 수정
