+++ src/main/java/subway/ui/exception/NotFoundException.java
@@ -1,0 +1,8 @@
+package subway.ui.exception;
+
+public abstract class NotFoundException extends RuntimeException {
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
